requirements.txt
flask
spacy
pdfplumber
python-docx
sentence-transformers
scikit-learn
pandas